+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - The function `GetFlightLog` to get a specific flight log of a user.
 - The function `GetFlightLogs` to get all of a user's flight logs.
+- The `GET /flihgt-log/-` endpoint handle the exchange to get a specific flight
+  log for a user.
+- The `GET /flight-log` endpoint to handle the exchange to get all of a user's
+  flight logs.
 
 ## [0.1.0] - 2022-12-10
 ### Added